==> notebooks/03_fine-tuning.ipynb <==
# importing libraries
import pandas as pd
import os
from dotenv import load_dotenv

# loading environment variables
load_dotenv()
DATA_PATH = os.getenv('DATA_PATH')

# loading data
df = pd.read_parquet(os.path.join(DATA_PATH, '../processed/dsi-qa-pairs.parquet'))
df.head()

from datasets import Dataset

# Format the data
def format_conversation(row):
    return [
        {"role": "system", "content": "You are an assistant."},
        {"role": "user", "content": row["question"]},
        {"role": "assistant", "content": row["answer"]},
    ]

df["conversations"] = df.apply(format_conversation, axis=1)

# Convert to Hugging Face Dataset
dataset = Dataset.from_pandas(df[["conversations"]])


dataset.save_to_disk(os.path.join(DATA_PATH, '../processed/dsi-qa-pairs-instruct.hf'))

### Using ChatGPT generated data


gpt_data = pd.read_parquet(os.path.join(DATA_PATH, '../processed/chat-gpt-qa-pairs.parquet'))
gpt_data.columns = ['topic', 'question', 'answer']
gpt_data.head()

# Apply the format_conversation function to each row in gpt_data to create a 'conversations' column
gpt_data['conversations'] = gpt_data.apply(format_conversation, axis=1)

# Convert the DataFrame to a Hugging Face Dataset object
gpt_dataset = Dataset.from_pandas(gpt_data[['conversations']])

# Save the dataset to disk
gpt_dataset.save_to_disk(os.path.join(DATA_PATH, '../processed/chat-gpt-qa-pairs-instruct.hf'))

What did the DSI-SRP fellowship grant Jingyi Chen to work in the laboratory of Christopher Candelaria?<br>
The AI model produces novel text that takes form of written correspondence, with style and content consistent with the data source?

gpt_dataset

# Split the dataset into train and test sets
split_datasets = gpt_dataset.train_test_split(test_size=0.2, seed=42)

# Access the train and test sets
train_dataset = split_datasets['train']
test_dataset = split_datasets['test']

# Display sizes for confirmation
print(f"Train dataset size: {len(train_dataset)}")
print(f"Test dataset size: {len(test_dataset)}")

train_dataset.save_to_disk(os.path.join(DATA_PATH, '../processed/chat-gpt-qa-pairs-instruct-train.hf'))
test_dataset.save_to_disk(os.path.join(DATA_PATH, '../processed/chat-gpt-qa-pairs-instruct-test.hf'))
